datepicker is now in own file
